Concatenation Of String
This C++ program illustrates how to concatenate strings using the + operator. String concatenation is the process of combining two or more strings to form a single string. In C++, the std::string class supports the + operator, making it easy to join strings together.

In this program, we cover:

Basic String Concatenation using the + operator.
Concatenating Multiple Strings in a single operation.
Storing the Result of Concatenation in a new string variable.