Left And Right Rotations In String
This C++ program demonstrates how to rotate a string either to the left or to the right by a given number of positions. A left rotation shifts each character in the string towards the left by one position, with the first character moving to the end. Similarly, a right rotation shifts the string to the right, with the last character moving to the front.

In this program, we cover:

Left Rotation of a string by a specified number of positions.
Right Rotation of a string by a specified number of positions.