Singly Linked List
This C++ program implements a Singly Linked List (SLL), a linear data structure where each element (node) points to the next node in the sequence. The last node points to nullptr, marking the end of the list. Unlike arrays, a linked list can dynamically grow and shrink, and elements can be added or removed efficiently.

In this program, we cover:

Node Creation for the linked list.
Inserting Elements at the beginning, end, and at a specified position.
Displaying the Linked List elements.
Deleting an Element from the list.