Reverse A linkd List
This C++ program demonstrates how to reverse a singly linked list. In a singly linked list, each node points to the next node, but to reverse the list, the direction of the pointers needs to be changed, so that each node points to its previous node. The last node will become the new head, and the previous head will point to nullptr, marking the new end of the list.

In this program, we cover:

Node Creation for the linked list.
Reversing the Linked List using an iterative method.
Displaying the Linked List before and after the reversal.