Deletion Of Head Element In Linked List
Description:
This C++ program demonstrates how to delete the head node of a singly linked list. The head node is the first node in the list, and deleting it requires updating the head pointer to the second node in the list. If the list only contains one node, the head will become nullptr, indicating that the list is now empty.

In this program, we cover:

Node Creation for the linked list.
Inserting Elements at the end of the list.
Deleting the Head Node.
Displaying the Linked List before and after the deletion.