Deletion at Specific Position of Linked List
Description:
This C++ program demonstrates how to delete a node at a specific position in a singly linked list. In a singly linked list, each node points to the next node. To delete a node at a given position, we need to update the next pointer of the previous node to bypass the node being deleted.

In this program, we cover:

Node Creation for the linked list.
Inserting Elements at the end of the list.
Deleting a Node at a Given Position.
Displaying the Linked List before and after the deletion.