Creating A Doubly Linked List
This C++ program demonstrates how to create a doubly linked list. A doubly linked list is a type of linked list where each node contains two pointers:

One pointing to the next node in the list.
One pointing to the previous node in the list.
This makes traversal of the list possible in both directions (forward and backward), unlike a singly linked list which can only be traversed in one direction.

In this program, we cover:

Node Creation for the doubly linked list.
Insertion of Elements at the end of the list.
Displaying the List in both forward and reverse order.