Doubly Linked List Insertion At Beginning
This C++ program demonstrates how to insert a new node at the beginning of a doubly linked list. In a doubly linked list, each node contains two pointers:

A pointer to the next node.
A pointer to the previous node.
The process for inserting at the beginning of the doubly linked list involves:

Creating a new node.
Adjusting the pointers of the current head and the new node.
Making the new node the head of the list.