Insertion At End
Description:
This C++ program demonstrates how to insert a new node at the end of a doubly linked list. In a doubly linked list, each node contains two pointers:

A pointer to the next node.
A pointer to the previous node.
The process for inserting at the end of a doubly linked list involves:

Creating a new node.
Traversing the list to the last node.
Adjusting the pointers of the last node and the new node to link them.